Add _SparseGLM class for more efficient fit of GLM models with sparse exogenous variables.
The sparsity of the exogenous variables is exploited in:
- collinearity checks, where the nullspace can be more quicky computed
  by a skinny qr decomposition of `exog.T.dot(exog).toarray()`
- computing the log-likelihood, score, and hessian in numerical
  optimization

Although exog is often very sparse, the Hessian is not,
so it is still returned as a dense array.

The method used to fit the GLM is switched from iteratively reweighted
least squares (IRLS) to a Newton linesearch method to maximize the
log-likelihood.  The advantage is avoiding dense QR-based solution of
least squares problems; the disadvantage is that the linear solves in
Newton's method look like solving weighted least squares problems by
invertin the normal equations matrix, which has twice the condition
number.  A compromise might be to conclude with a single dense
QR-based step if desired.

Add SparseGLM and _fit_newton_line_search

Strip away development cruft

Remove development files