Allow more accurate WLS at conclusion of _fit_newton_line_search
When solving the update equations in Newton's method by inverting the
normal equations matrix, the condition number of the problem is squared
and a less accurate solution is obtained.  In this update, if the
line search fails using this method, the Newton update can be computed
by solving the equivalent weighted least squares problem, giving a more
accurate update direction.